chore(gRPC): move `CheckpointTransactionWithContext` in `iota-grpc-server`
